Refactor main loop to cleaner code
This is a multipart refactor commit to clean up and refactor the BuildGuestBook method main loop. The initial greeting before the main loop was refactored into its own method. Secondly, a do-while loop when getting the party size was refactored to instead use a goto statement when we need to repeat the process of getting a party size. Third, for clarity, extensive commenting documenting how the loop works was added.